+++ src/sheets/views/ui_views.py
@@ -2,7 +2,7 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse_lazy
 from django.views.generic import TemplateView
 
-from accounts.models import ALL_USER_CATEGORIES
+from accounts.constants import ALL_USER_CATEGORIES
 from common.mixins import FullyLoggedMixin
 from content.models import FeedbackResult
 
@@ -12,7 +12,7 @@ class PublicHomeView(TemplateView):
 
     def get(self, request, *args, **kwargs):
         """Redirect user to private home or second_factor page wether they're logged in or verified."""
-        if request.user.is_verified() or request.user.is_authenticated_from_monaiot():
+        if request.user.is_verified() or request.user.is_authenticated_from_oidc():
             return HttpResponseRedirect(reverse_lazy("private_home"))
         if request.user.is_authenticated:
             return HttpResponseRedirect(reverse_lazy("second_factor"))