+++ changelog.md
@@ -5,6 +5,10 @@ Les changements importants de Trackdéchets préparation inspection sont documen
 Le format est basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 et le projet suit un schéma de versioning inspiré de [Calendar Versioning](https://calver.org/).
 
+## 25/02/2025
+
+- Modification des permissions d'accès de l'administration centrale
+
 ## 24/02/2025
 
 - Utilisation de mozilla-oidc pour la connexion MonAIOT
@@ -32,7 +36,8 @@ et le projet suit un schéma de versioning inspiré de [Calendar Versioning](htt
 ## 01/01/2025
 
 - Correction de la recherche par plaque
-
+BSD-20250225-TEWGBTTVD
+- 
 ## 26/11/2024
 
 - Ajout de la cartographie des établissements